Add higher resolution vbat
The excellent vbat compensation that was recently added is somewhat
hampered by the limited resolution of the current filtered battery
voltage (0.1V). This means there are typically only around 10 steps of
compensation available during the flight, and if flying smoothly each
step can be seen to cause an upward bounce as the throttle increases.
This change adds a new hires (0.01V) filtered vbat field and an
independent filter to go with it. For my own use, I also changed the OSD
display of average cell voltage to use the higher resolution value.